+++ packages/style-kit/eslint.config.js
@@ -4,4 +4,8 @@ export default eslintConfig({
   jsdoc: {
     requireJsdoc: true,
   },
+  testing: {
+    framework: "bun",
+  },
+  typescript: "tsconfig.eslint.json",
 });