+++ style-kit.config.js
@@ -1,0 +1,7 @@
+import { eslintConfig, prettierConfig } from "js-style-kit";
+
+export const prettier = prettierConfig();
+
+export const eslint = eslintConfig({
+  turbo: true,
+});